This commit enhances ssh-related operations:
- fast-vm keydist <VmNumber|'VmNumber1 VmNumber2 ...'>
  Distribute the fast-vm ssh key (using ssh-copy-id) to one or more VMs

keydist will automatically generate a passphraseless ssh key for
fast-vm's use, if it doesn't already exist, and store it in
$FASTVM_USER_CONF_DIR

If found, it will be used automatically for ssh and scp.  Can be run
against multiple VMs at once and will iterate through each in turn.

- fast-vm scp <VmNumber|'VmNumber1 VmNumber2 ...'> <vm:/remote/sourcefile /local/sourcefile|/local/destfile vm:/remote/destfile>

Provide an easy way to scp files between the host and guest VMs,
using the ssh key if it exists.  "vm" as a hostname in the source
or destination will be replaced with the VM addresses specified,
allowing easy copy-to-vm or copy-from-vm interaction.

- fast-vm ssh changes to utilize the fast-vm ssh key if it exists